requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
werkzeug==3.1.3
email-validator==2.2.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
